fehler in header datei gefixt

+++ game.h
@@ -2,6 +2,6 @@
 #define GAME_H
 
 void game_init();
-void game_tick(double delta);
+void game_tick(int delta);
 
 #endif
